V0.0.4
Big Update!

MAD-AS_Sampler:
The original MAD-AS program is now written into the code and the entire code is cleaned up!
With adding the remote control function to the program, there are now three options to activate sampling operation.
The capability to record the number of cycles/activation/power by using EEPROM is temporarily removed.
Program tested on MAD-AS.

MAD-AS_Controller:
The code is cleaned up. No additional function yet.