V0.0.6 (a stable version)
This should be a stable version (comparing to V0.0.5) which realises the function to wake up the MAD-AS Sampler after the Controller receives "YES" from the cloud. In the previous version, many problems appeared in the testing. However, more testing is still needed in case this version is still problematic in some situations.

Fixed the following bugs:
1. MAD-AS Controller could not hear the response from MAD-AS Sampler. This is due to the conflict between two software serial (simCom/HC12).
2. IF loop condition could be violated. This is because the condition variable is not well defined.