V0.0.7 (Big Update with Cycle Counter Added)
Cycle Counter function is now added in this HUGE update!

EEPROM is added back to the sampler program, so the sampler is now able to record the number of cycles it performs. In addition to this, the number of cycle can be sent to the controller and then sent to the cloud after a new sampling cycle is performed! By this way, we are able to understand the real-time status of MAD-AS sampler on our website!