V0.0.8 (stop command and time-out functions)
Finally we got some mechanism to stop the programs and let the microcontroller sleep!

Stop Sampling Function: when the "stop sampling" command is pushed on the web and the controller finds the command, the controller will try to communicate and stop the sampler to take any further. Due to this function, the sampler will turn on radio for 4 secs after being triggered by the alarm, before taking new sample.

Time-out Function: if the controller does not hear from the sampler for more than 2 hours, it will sleep forever on its own.

Before the controller sleeps forever, it will send one last message to the web informing the reason why it turns off.